paypal: remove formatters class
